Server/requirement.txt
flask==2.2.2
flask-mysql==1.5.2
flask-cors==3.0.10
Flask-Mail==0.9.1
python-dotenv==0.21.0
Flask-Reuploaded==1.3.0